[#1289] Move adapters
