+++ streampipes-extensions/streampipes-connect-adapters-iiot/src/main/java/org/apache/streampipes/connect/iiot/ConnectAdapterIiotInit.java
@@ -22,8 +22,8 @@ import org.apache.streampipes.connect.iiot.adapters.opcua.OpcUaAdapter;
 import org.apache.streampipes.connect.iiot.adapters.plc4x.modbus.Plc4xModbusAdapter;
 import org.apache.streampipes.connect.iiot.adapters.plc4x.s7.Plc4xS7Adapter;
 import org.apache.streampipes.connect.iiot.adapters.ros.RosBridgeAdapter;
-import org.apache.streampipes.connect.iiot.adapters.simulator.machine.v2.FileReplayAdapter;
-import org.apache.streampipes.connect.iiot.adapters.simulator.machine.v2.NewMachineDataSimulatorAdapter;
+import org.apache.streampipes.connect.iiot.adapters.simulator.machine.MachineDataSimulatorAdapter;
+import org.apache.streampipes.connect.iiot.protocol.stream.FileReplayAdapter;
 import org.apache.streampipes.connect.iiot.protocol.stream.HttpServerProtocol;
 import org.apache.streampipes.connect.iiot.protocol.stream.HttpStreamProtocol;
 import org.apache.streampipes.connect.iiot.protocol.stream.KafkaProtocol;
@@ -47,14 +47,12 @@ public class ConnectAdapterIiotInit extends ExtensionsModelSubmitter {
             "StreamPipes connect worker containing adapters relevant for the IIoT",
             "",
             8001)
-        .registerAdapter(new NewMachineDataSimulatorAdapter())
+        .registerAdapter(new MachineDataSimulatorAdapter())
         .registerAdapter(new FileReplayAdapter())
-//        .registerAdapter(new MachineDataStreamAdapter())
         .registerAdapter(new RosBridgeAdapter())
         .registerAdapter(new OpcUaAdapter())
         .registerAdapter(new Plc4xS7Adapter())
         .registerAdapter(new Plc4xModbusAdapter())
-//        .registerAdapter(new FileStreamProtocol())
         .registerAdapter(new KafkaProtocol())
         .registerAdapter(new MqttProtocol())
         .registerAdapter(new NatsProtocol())
